requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
pandas==2.1.1
numpy==1.24.3
python-dotenv==1.0.0

# AI/ML dependencies
tensorflow==2.13.0
scikit-learn==1.3.0
keras==2.13.1

# Technical Analysis
ta==0.10.2
yfinance==0.2.18
ccxt==4.0.77

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.21

# Social Media & News
tweepy==4.14.0
praw==7.7.1
feedparser==6.0.10
newspaper3k==0.2.8
textblob==0.17.1
nltk==3.8.1

# Telegram
python-telegram-bot==20.4

# Data processing
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.0

# Utilities
schedule==1.2.0
APScheduler==3.10.4
python-dateutil==2.8.2
pytz==2023.3

# Web scraping
aiohttp==3.8.6
asyncio==3.4.3

# JSON handling
ujson==5.8.0

# Logging
colorlog==6.7.0 